chore(federation): move normalize() implementations to a module
