+++ apollo-federation/src/operation/optimize.rs
@@ -53,7 +53,6 @@ use super::Fragment;
 use super::FragmentSpreadSelection;
 use super::InlineFragmentSelection;
 use super::NamedFragments;
-use super::NormalizeSelectionOption;
 use super::Operation;
 use super::Selection;
 use super::SelectionKey;
@@ -82,7 +81,7 @@ impl NamedFragments {
                 &fragment.type_condition_position,
                 &Default::default(),
                 &fragment.schema,
-                NormalizeSelectionOption::NormalizeRecursively,
+                Default::default(),
             )?;
             let mut mapped_selection_set = mapper(&expanded_selection_set)?;
             // `mapped_selection_set` must be fragment-spread-free.
@@ -650,7 +649,7 @@ impl Fragment {
             ty,
             /*named_fragments*/ &Default::default(),
             &self.schema,
-            NormalizeSelectionOption::NormalizeRecursively,
+            Default::default(),
         )?;
 
         if !self.type_condition_position.is_object_type() {
@@ -1165,7 +1164,7 @@ impl NamedFragments {
                     &fragment.selection_set.type_position,
                     &fragments_to_keep,
                     &fragment.schema,
-                    NormalizeSelectionOption::NormalizeRecursively,
+                    Default::default(),
                 )?;
         }
 
@@ -1184,7 +1183,7 @@ impl NamedFragments {
             &reduced_selection_set.type_position,
             self,
             &selection_set.schema,
-            NormalizeSelectionOption::NormalizeRecursively,
+            Default::default(),
         )
     }
 
@@ -1476,7 +1475,7 @@ impl Operation {
             &self.selection_set.type_position,
             &self.named_fragments,
             &self.schema,
-            NormalizeSelectionOption::NormalizeRecursively,
+            Default::default(),
         )?;
         Ok(Self {
             named_fragments: Default::default(),
